+++ src/main/java/insane96mcp/enhancedai/modules/mobs/flee/Flee.java
@@ -6,6 +6,7 @@ import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.JsonFeature;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.data.IdTagMatcher;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
@@ -16,7 +17,7 @@ import java.util.List;
 
 @LoadFeature(module = Modules.Ids.MOBS, canBeDisabled = false, description = "Custom Json config to set mobs from running from other mobs")
 public class Flee extends JsonFeature {
-	public static final List<CustomFleeConfig> CUSTOM_FLEE_DEFAULT = List.of();
+	public static final List<CustomFleeConfig> CUSTOM_FLEE_DEFAULT = List.of(new CustomFleeConfig(1, IdTagMatcher.newTag("enhancedai:all_mobs"), IdTagMatcher.newId("minecraft:warden"), 1d, 16, 8, 1, 1.1));
 
 	public static final List<CustomFleeConfig> customFlee = new ArrayList<>();