+++ changelog.md
@@ -22,6 +22,8 @@ The tags have been moved to their folders (e.g. `enhancedai:can_be_miner` -> `en
   * Drowned have 30 seconds where they don't burn in sunlight. During this time, they will not seek water
 * Miner mobs now use the `forge:block_reach` attribute
 * Melee mobs default attack speed multiplier is now 0.25 (so 1 attack/sec instead of 2)
+* Removed Warden Interactions
+  * Fleeing is done through Custom Flee feature
 * Pillagers now try to reposition to not hit allies (or when they hit an ally)
 * Merged Pillager and Witch flee features into a single "Target Flee" feature
 * Web throwing feature is no longer limited to spiders